processing code moved to imageFilter routes
